+++ src/modules/aws_api_gateway/mappers/index.ts
@@ -1,0 +1,1 @@
+export * from './api';
